# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JDS Platform API",
        "description": "API documentation for the JDS Platform Laravel backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/exco/member-applications/{id}": {
            "get": {
                "tags": [
                    "EXCO Member Applications"
                ],
                "summary": "View a single member application",
                "operationId": "93015de0ec9489bad392190f9efa6ee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberApplication"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/exco/members/{member}": {
            "get": {
                "tags": [
                    "EXCO Members"
                ],
                "summary": "View a single member for EXCO",
                "operationId": "d8a0dcfa3a8d0e814055f873dc96747f",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Member": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "member_number": {
                        "type": "string",
                        "example": "JDS2026-0001",
                        "nullable": true
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08012345678"
                    },
                    "joined_on": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-04-01",
                        "nullable": true
                    },
                    "membership_status": {
                        "type": "string",
                        "example": "active"
                    },
                    "has_online_banking": {
                        "type": "boolean",
                        "example": true
                    },
                    "whatsapp_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "profile_completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T08:23:06Z",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Created from member application #3",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    },
                    "share_purchases_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "loans_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "concerns_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "share_purchases_sum_shares_count": {
                        "type": "number",
                        "format": "float",
                        "example": 8,
                        "nullable": true
                    },
                    "share_purchases_sum_total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 160000,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T08:23:06Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T08:23:06Z"
                    }
                },
                "type": "object"
            },
            "MemberApplication": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "membership_cycle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "jane@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08012345678"
                    },
                    "has_online_banking": {
                        "type": "boolean",
                        "example": true
                    },
                    "whatsapp_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "biodata": {
                        "type": "object",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "pending_review"
                    },
                    "applied_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T20:10:00Z",
                        "nullable": true
                    },
                    "reviewed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "reviewed_by": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "approved_member_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "review_note": {
                        "type": "string",
                        "nullable": true
                    },
                    "cycle": {
                        "$ref": "#/components/schemas/MembershipCycle"
                    },
                    "reviewer": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    },
                    "approved_member": {
                        "$ref": "#/components/schemas/Member"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T20:10:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T20:10:00Z"
                    }
                },
                "type": "object"
            },
            "MembershipCycle": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "JDS 2026 Cycle"
                    },
                    "code": {
                        "type": "string",
                        "example": "JDS2026"
                    },
                    "starts_on": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01"
                    },
                    "ends_on": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31"
                    },
                    "onboarding_opens_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T00:00:00Z",
                        "nullable": true
                    },
                    "onboarding_closes_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-31T23:59:59Z",
                        "nullable": true
                    },
                    "accepting_new_applications": {
                        "type": "boolean",
                        "example": true
                    },
                    "onboarding_notes": {
                        "type": "string",
                        "example": "Initial demo cycle for local development.",
                        "nullable": true
                    },
                    "share_price": {
                        "type": "number",
                        "format": "float",
                        "example": 20000
                    },
                    "min_monthly_shares": {
                        "type": "integer",
                        "example": 1
                    },
                    "max_monthly_shares": {
                        "type": "integer",
                        "example": 10
                    },
                    "loan_multiplier": {
                        "type": "number",
                        "format": "float",
                        "example": 3
                    },
                    "loan_service_charge_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "loan_duration_months": {
                        "type": "integer",
                        "example": 3
                    },
                    "overdue_penalty_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "overdue_penalty_window_months": {
                        "type": "integer",
                        "example": 2
                    },
                    "shareout_admin_fee_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "registration_fee_new_member": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "registration_fee_existing_member": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "loan_pause_month": {
                        "type": "integer",
                        "example": 9
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "member_applications_count": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "share_purchases_count": {
                        "type": "integer",
                        "example": 47,
                        "nullable": true
                    },
                    "loans_count": {
                        "type": "integer",
                        "example": 9,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T19:26:17Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T19:26:17Z"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Shadrach Gideon"
                    },
                    "email": {
                        "type": "string",
                        "example": "chairperson@jds.local"
                    },
                    "member_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "chairperson"
                    },
                    "must_change_password": {
                        "type": "boolean",
                        "example": true
                    },
                    "temporary_password_sent_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-07T09:00:00Z",
                        "nullable": true
                    },
                    "password_changed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-07T09:30:00Z",
                        "nullable": true
                    },
                    "member": {
                        "$ref": "#/components/schemas/Member"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T19:26:18Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-01T19:26:18Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the Bearer token returned from POST /api/auth/login.",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}